Support decimals for SSL_times. No longer using thread-local timing.
svn path=/spoofax/trunk/spoofax/org.spoofax.interpreter.core/; revision=19697

+++ org.spoofax.interpreter.core/src/java/org/spoofax/interpreter/library/ssl/SSL_TicksToSeconds.java
@@ -31,7 +31,7 @@ public class SSL_TicksToSeconds extends AbstractPrimitive {
         else
             return false;
         
-        env.setCurrent(env.getFactory().makeReal(timeValue / SSL_times.TICKS_PER_SECOND));
+        env.setCurrent(env.getFactory().makeReal(timeValue / (double) SSL_times.TICKS_PER_SECOND));
         return true;
     }
